Improve enum class functionality across repo
• Moved bitfield into dedicated template, separate templates for legacy enums & enum classes
• New specialized defines for binding enum classes, adding bitwise operators
• Variant enum class define block changed to conditional template; template added for bitfield as well
• Enum class BitFields now expect fully qualified enum values, but those values are now much easier to perform bitwise operations on

+++ core/input/input_event.cpp
@@ -385,11 +385,11 @@ bool InputEventKey::is_echo() const {
 }
 
 Key InputEventKey::get_keycode_with_modifiers() const {
-	return keycode | (int64_t)get_modifiers_mask();
+	return keycode | get_modifiers_mask();
 }
 
 Key InputEventKey::get_physical_keycode_with_modifiers() const {
-	return physical_keycode | (int64_t)get_modifiers_mask();
+	return physical_keycode | get_modifiers_mask();
 }
 
 Key InputEventKey::get_key_label_with_modifiers() const {
